+++ CHANGELOG.md
@@ -6,7 +6,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
-
+- Added `discover_combined_schema` function in Polars Utils to generate combined schema from a partitioned dataset.
 
 ### Changed
 - Pull clean workplace columns into merge job within SLV pipeline.